[internal] Minor refacto, renaming and moving files for clarity
