+++ src/pages/api/lesson/index.ts
@@ -1,6 +1,6 @@
 import { NextApiHandler, NextApiRequest, NextApiResponse } from "next";
 import { lessonGetAllHandler } from "@handlers/lesson/get.handler";
-import { lessonPostHandler } from "@handlers/lesson/post.handler";
+import { lessonPostHandler } from "@handlers/lesson/create.handler";
 import routerMiddleware from "@middlewares/router.middleware";
 import { ApiResponse } from "@typing/api-response.interface";
