+++ src/server/handlers/user/login/auto-login.handler.ts
@@ -1,5 +1,4 @@
 import { NextApiHandler, NextApiRequest, NextApiResponse } from "next";
-import { withSession } from "@middlewares/session.middleware";
 import { ApiResponse } from "@typing/api-response.interface";
 import { IUserPublic } from "@typing/user.interface";
 
@@ -22,4 +21,4 @@ const handler = async (
   });
 };
 
-export const autoLoginHandler: NextApiHandler = withSession(handler);
+export const autoLoginHandler: NextApiHandler = handler;
