+++ src/server/middlewares/authentication/auth-check.middleware.ts
