+++ src/server/middlewares/authentication/session.middleware.ts
