+++ src/server/middlewares/router.middleware.ts
@@ -1,7 +1,7 @@
 import { NextApiHandler } from "next";
 import { notImplementedHandler } from "@common/not-implemented.handler";
-import authCheckMiddleware from "@middlewares/auth-check.middleware";
-import { withSession } from "@middlewares/session.middleware";
+import authCheckMiddleware from "@middlewares/authentication/auth-check.middleware";
+import { withSession } from "@middlewares/authentication/session.middleware";
 import { ApiResponse } from "@typing/api-response.interface";
 
 /**
@@ -31,6 +31,6 @@ export default (
   if (requiresAuth) {
     return withSession(authCheckMiddleware(handler));
   } else {
-    return handler;
+    return withSession(handler);
   }
 };
